Move code out of WorkerChannel and setupCoreModule
No functional changes, just moving code
