+++ src/hooks/executeHooks.ts
@@ -1,0 +1,37 @@
+// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the MIT License.
+
+import { HookContext } from '@azure/functions-core';
+import { AzureFunctionsRpcMessages as rpc } from '../../azure-functions-language-worker-protobuf/src/rpc';
+import { channel } from '../WorkerChannel';
+import { getHooks } from './getHooks';
+import LogLevel = rpc.RpcLog.Level;
+import LogCategory = rpc.RpcLog.RpcLogCategory;
+
+export async function executeHooks(
+    hookName: string,
+    context: HookContext,
+    invocationId?: string | null,
+    msgCategory?: string
+): Promise<void> {
+    const callbacks = getHooks(hookName);
+    if (callbacks.length > 0) {
+        channel.log({
+            message: `Executing ${callbacks.length} "${hookName}" hooks`,
+            level: LogLevel.Debug,
+            logCategory: LogCategory.System,
+            invocationId,
+            category: msgCategory,
+        });
+        for (const callback of callbacks) {
+            await callback(context);
+        }
+        channel.log({
+            message: `Executed "${hookName}" hooks`,
+            level: LogLevel.Debug,
+            logCategory: LogCategory.System,
+            invocationId,
+            category: msgCategory,
+        });
+    }
+}
