+++ src/setupCoreModule.ts
@@ -1,15 +1,13 @@
 // Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the MIT License.
 
-import { FunctionCallback, FunctionMetadata, HookCallback, ProgrammingModel } from '@azure/functions-core';
-import { AzureFunctionsRpcMessages as rpc } from '../azure-functions-language-worker-protobuf/src/rpc';
 import { Disposable } from './Disposable';
 import { channel } from './WorkerChannel';
 import { version } from './constants';
 import { registerFunction } from './coreApi/registerFunction';
+import { setProgrammingModel } from './coreApi/setProgrammingModel';
+import { registerHook } from './hooks/registerHook';
 import Module = require('module');
-import LogCategory = rpc.RpcLog.RpcLogCategory;
-import LogLevel = rpc.RpcLog.Level;
 
 /**
  * Intercepts the default "require" method so that we can provide our own "built-in" module
@@ -22,26 +20,12 @@ export function setupCoreModule(): void {
         get hostVersion() {
             return channel.hostVersion;
         },
-        registerHook: (hookName: string, callback: HookCallback) => channel.registerHook(hookName, callback),
-        setProgrammingModel: (programmingModel: ProgrammingModel) => {
-            // Log when setting the programming model, except for the initial default one (partially because the grpc channels aren't fully setup at that time)
-            if (channel.app.programmingModel) {
-                channel.log({
-                    message: `Setting Node.js programming model to "${programmingModel.name}" version "${programmingModel.version}"`,
-                    level: LogLevel.Information,
-                    logCategory: LogCategory.System,
-                });
-            } else {
-                channel.defaultProgrammingModel = programmingModel;
-            }
-            channel.app.programmingModel = programmingModel;
-        },
+        registerHook,
+        setProgrammingModel,
         getProgrammingModel: () => {
             return channel.app.programmingModel;
         },
-        registerFunction: (metadata: FunctionMetadata, callback: FunctionCallback) => {
-            return registerFunction(metadata, callback);
-        },
+        registerFunction,
         Disposable,
     };
